refactor : FridgeHomeScreen 컴포넌트 분리
