+++ fresco_frontend/src/components/FridgeHome/FridgeItemCard/styles.ts
@@ -1,0 +1,87 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  itemCard: {
+    flexDirection: 'row',
+    backgroundColor: '#fff',
+    borderRadius: 12,
+    padding: 16,
+    marginBottom: 12,
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+  },
+  itemImageContainer: {
+    marginRight: 16,
+  },
+  itemImagePlaceholder: {
+    width: 60,
+    height: 60,
+    backgroundColor: '#e0e0e0',
+    borderRadius: 8,
+  },
+  itemInfo: {
+    flex: 1,
+    justifyContent: 'space-between',
+  },
+  itemName: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    color: '#333',
+    marginBottom: 4,
+  },
+  itemDetails: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginBottom: 4,
+  },
+  itemQuantity: {
+    fontSize: 14,
+    color: '#666',
+    marginRight: 12,
+  },
+  itemExpiry: {
+    fontSize: 14,
+    color: '#666',
+  },
+  itemStatus: {
+    fontSize: 12,
+    color: '#999',
+    fontStyle: 'italic',
+  },
+  // 편집 모드용 스타일
+  editModeActions: {
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    marginLeft: 8,
+    gap: 4,
+  },
+  editButton: {
+    backgroundColor: '#007AFF',
+    paddingHorizontal: 8,
+    paddingVertical: 4,
+    borderRadius: 4,
+  },
+  editButtonText: {
+    fontSize: 12,
+    color: '#fff',
+    fontWeight: '500',
+  },
+  deleteButton: {
+    backgroundColor: '#FF3B30',
+    paddingHorizontal: 8,
+    paddingVertical: 4,
+    borderRadius: 4,
+  },
+  deleteButtonText: {
+    fontSize: 12,
+    color: '#fff',
+    fontWeight: '500',
+  },
+});
