+++ fresco_frontend/src/components/FridgeSelect/FridgeTile/styles.ts
