+++ fresco_frontend/src/hooks/useModalState.ts
@@ -1,0 +1,22 @@
+import {useState} from 'react';
+
+export const useModalState = () => {
+  const [isStorageModalVisible, setIsStorageModalVisible] = useState(false);
+  const [isItemCategoryModalVisible, setIsItemCategoryModalVisible] =
+    useState(false);
+
+  const openStorageModal = () => setIsStorageModalVisible(true);
+  const closeStorageModal = () => setIsStorageModalVisible(false);
+
+  const openItemCategoryModal = () => setIsItemCategoryModalVisible(true);
+  const closeItemCategoryModal = () => setIsItemCategoryModalVisible(false);
+
+  return {
+    isStorageModalVisible,
+    isItemCategoryModalVisible,
+    openStorageModal,
+    closeStorageModal,
+    openItemCategoryModal,
+    closeItemCategoryModal,
+  };
+};
